move storage.ts to lib
